Projects/estoque/backend/requirements.txt
Flask
flask-login
Flask-CORS
psycopg2
python-dotenv
